requirements-dev.txt
-r requirements.txt
coverage
doc8
flake8
mock
pylint
pytest
pytest-runner
pytest-mock
pyyaml
Sphinx
sphinxcontrib-napoleon
